Splite brixelate addon into multiple files
