+++ test.py
@@ -1,0 +1,9 @@
+def testfunc():
+
+	print("Test func working")
+
+
+class TestClass():
+
+	def classmethod():
+		print("Test class workgin")
